chore: mv settings and generator to lib
