+++ offchain-modules/bridge-lib/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+sdk = { path = "../sdk" }
 anyhow = "1.0"
 ethereum-types = {version = "0.9.2" }
 web3 = "0.8.0"
@@ -19,4 +20,7 @@ ckb-types = { git = "https://github.com/nervosnetwork/ckb", tag = "v0.36.0-rc2" 
 ckb-jsonrpc-types = { git = "https://github.com/nervosnetwork/ckb", tag = "v0.36.0-rc2" }
 log = "0.4"
 serde_json = "1.0"
-sdk = { path = "../sdk" }
+serde = { version = "1.0", features = ["rc"] }
+serde_derive = "1.0"
+config = "0.10"
+toml = "0.5"
