--- offchain-modules/bridge-lib/src/util/generator.rs
@@ -1,100 +1,0 @@
-use ckb_sdk::{
-    GenesisInfo, HttpRpcClient,
-};
-use ckb_types::prelude::Pack;
-use ckb_types::{
-    bytes::Bytes,
-    core::{BlockView, DepType},
-    packed::{self, Byte32, CellDep, CellOutput, OutPoint, Script},
-    prelude::{Builder, Entity},
-};
-use sdk::indexer::IndexerRpcClient;
-use sdk::settings::{Settings, OutpointConf};
-use sdk::tx_helper::TxHelper;
-use sdk::cell_collector::get_live_cell_by_typescript;
-use sdk::util::get_live_cell;
-
-
-pub struct Generator {
-    pub rpc_client: HttpRpcClient,
-    pub indexer_client: IndexerRpcClient,
-    _genesis_info: GenesisInfo,
-    _settings: Settings,
-}
-
-impl Generator {
-    pub fn new(rpc_url: String, indexer_url: String, settings: Settings) -> Result<Self, String> {
-        let mut rpc_client = HttpRpcClient::new(rpc_url);
-        let indexer_client = IndexerRpcClient::new(indexer_url);
-        let genesis_block: BlockView = rpc_client
-            .get_block_by_number(0)?
-            .expect("Can not get genesis block?")
-            .into();
-        let genesis_info = GenesisInfo::from_block(&genesis_block)?;
-        Ok(Self {
-            rpc_client,
-            indexer_client,
-            _genesis_info: genesis_info,
-            _settings: settings,
-        })
-    }
-
-
-
-    fn _add_cell_deps(
-        &mut self,
-        helper: &mut TxHelper,
-        outpoints: Vec<OutpointConf>,
-    ) -> Result<(), String> {
-        let mut builder = helper.transaction.as_advanced_builder();
-        for conf in outpoints {
-            let outpoint = OutPoint::new_builder()
-                .tx_hash(
-                    Byte32::from_slice(
-                        &hex::decode(conf.tx_hash)
-                            .map_err(|e| format!("invalid OutpointConf config. err: {}", e))?,
-                    )
-                    .map_err(|e| format!("invalid OutpointConf config. err: {}", e))?,
-                )
-                .index(conf.index.pack())
-                .build();
-            builder = builder.cell_dep(
-                CellDep::new_builder()
-                    .out_point(outpoint)
-                    .dep_type(DepType::Code.into())
-                    .build(),
-            );
-        }
-        helper.transaction = builder.build();
-        Ok(())
-    }
-
-
-
-    fn _get_ckb_cell(
-        &mut self,
-        helper: &mut TxHelper,
-        cell_typescript: Script,
-        add_to_input: bool,
-    ) -> Result<(CellOutput, Bytes), String> {
-        let genesis_info = self._genesis_info.clone();
-        let cell = get_live_cell_by_typescript(&mut self.indexer_client, cell_typescript)?
-            .ok_or(format!("cell not found"))?;
-        let ckb_cell = CellOutput::from(cell.output);
-        let ckb_cell_data = packed::Bytes::from(cell.output_data).raw_data();
-        if add_to_input {
-            let mut get_live_cell_fn = |out_point: OutPoint, with_data: bool| {
-                get_live_cell(&mut self.rpc_client, out_point, with_data).map(|(output, _)| output)
-            };
-
-            helper.add_input(
-                cell.out_point.into(),
-                None,
-                &mut get_live_cell_fn,
-                &genesis_info,
-                true,
-            )?;
-        }
-        Ok((ckb_cell, ckb_cell_data))
-    }
-}
