+++ offchain-modules/bridge-lib/src/util/mod.rs
@@ -1,3 +1,3 @@
 pub mod ckb_util;
 pub mod eth_util;
-pub mod generator;
+pub mod settings;
