+++ offchain-modules/sdk/Cargo.toml
@@ -29,7 +29,3 @@ serde_json = "1.0"
 
 [dev-dependencies]
 env_logger = "0.7.1"
-
-[lib]
-name = "sdk"
-path = "src/lib.rs"
