+++ offchain-modules/sdk/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod cell_collector;
 pub mod indexer;
-pub mod settings;
 pub mod tx_helper;
 pub mod util;
 pub mod constants;
